Removed duplicate classes and consolidated tests
